chore: converted all ID variables within models and usages to be long

+++ src/main/java/com/malabudi/dineupbe/menu/exception/MenuGroupNotFoundException.java
@@ -11,7 +11,7 @@ public class MenuGroupNotFoundException extends ResourceNotFoundException {
         super(message);
     }
 
-  public MenuGroupNotFoundException(Integer id ) {
+  public MenuGroupNotFoundException(Long id ) {
     super("Menu group with id " + id + " not found");
   }
 }
